requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain>=0.1.16
openai>=1.25.0
google-api-python-client>=2.127.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
aiohttp>=3.9.5
python-dotenv>=1.0.1
pydantic>=2.7.1
streamlit>=1.34.0